refactor: convert code to ES6
Side effects:

STATE is no longer exported. This is not a "breaking change" properly
speaking because it was never advertized formally as part of the API.

BREAKING CHANGE:

It is no longer possible to load the library as-is through a
``script`` element. It needs building.

The library now assumes a modern runtime. It no longer contains any
code to polyfill what's missing. It is up to developers using this
code to deal with polyfills as needed.