perf: simplify captureWhile
Notably the atStart parameter is no longer computed and passed to the
test. Those places that need it need to compute it themselves.

In fact most places that used it don't need it. The ``isNameStartChar`` checks
were mostly redundant as the necessary check on the first character of names was
done elsewhere.