refactor: remove startTagPosition
`startTagPosition` is part of the API of sax but was removed from the API when
saxes was forked. Its only remaining use was to produce padding when running
into an error case. This would allow producing a more representative text node
but accurrate data after an error is not part of saxes' contract. So the padding
has been removed and `startTagPosition` too.