refactor: rename textNode to text and use text for attribValue
The `attribValue` has a lifetime which never overlaps with `text` (formerly
`textNode`) so there's nothing gained by saving attribute values in
`attribValue` rather than `text`. So from now on, we also use `text` for
attribute values.

This allows simplifying the entity processing code because it can always put the
result in `text` instead of having to keep track of which buffer needs the
parsed entity.