feat: saxes handles chunks that "break" unicode
The way JavaScript handles unicode it is possible for a unicode string to be
split in two such that a unicode code point is split into two parts: e.g. ``
"\u{1F4A9}"[0]; "\u{1F4A9}"[1]`` If a large piece of data was cut up into
smaller chunks that were fed in sequence to saxes, and it so happened that a
unicode character was chopped up like illustrated above, then saxes would just
raise an error and fail. Saxes now detects when a chunk ends with a surrogate
and carries it over to the next chunk.