feat: formal method for setting event listeners
saxes now require the use of formal methods for setting event listeners.

BREAKING CHANGE: The individually named event handlers no longer exist. You now
must use the methods `on` and `off` to set handlers. Upcoming features require
that saxes know when handlers are added and removed, and it may be necessary in
the future to qualify how to add or remove a handler. Getters/setters are too
restrictives so we bite the bullet now and move to actual methods.